Beginning of "level 2": tagging bytes with metadata around what field of an x86 instruction they're for.
The x86 instruction set is variable-length, and how a byte is interpreted
affects later instruction boundaries. A lot of the pain in programming
machine code stems from computer and programmer going out of sync on what
a byte means. The miscommunication is usually not immediately caught, and
metastasizes at runtime into kilobytes of misinterpreted instructions.

To mitigate these issues, we'll start programming in terms of logical
operands rather than physical bytes. Some operands are smaller than a
byte, and others may consist of multiple bytes. This layer will correctly
pack and order the bytes corresponding to the operands in an instruction.